# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.GetPurchaseReportResponse:
    properties:
      data:
        $ref: '#/definitions/services.PurchaseReportData'
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.APIResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      status:
        example: success
        type: string
    type: object
...
  services.BalanceDiscrepancy:
    properties:
      cash_bank_balance:
        type: number
      cash_bank_code:
        type: string
      cash_bank_id:
        type: integer
      cash_bank_name:
        type: string
      detected_at:
        type: string
      difference:
        type: number
      gl_account_id:
        type: integer
      gl_balance:
        type: number
      gl_code:
        type: string
      gl_name:
        type: string
    type: object
  services.BalanceMonitoringResult:
    properties:
      check_time:
        type: string
      discrepancies:
        items:
          $ref: '#/definitions/services.BalanceDiscrepancy'
        type: array
      status:
        description: '"OK", "WARNING", "ERROR"'
        type: string
      synchronized_accounts:
        type: integer
      total_accounts_checked:
        type: integer
      unsynchronized_accounts:
        type: integer
    type: object
...
  services.CategoryPurchaseSummary:
    properties:
      account_code:
        type: string
      account_name:
        type: string
      category_name:
        type: string
      percentage:
        type: number
      total_amount:
        type: number
      total_purchases:
        type: integer
    type: object
  services.CompanyInfo:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tax_number:
        type: string
      website:
        type: string
    type: object
...
  services.MonthlyPurchaseSummary:
    properties:
      average_amount:
        type: number
      month:
        type: integer
      month_name:
        type: string
      total_amount:
        type: number
      total_paid:
        type: number
      total_purchases:
        type: integer
      year:
        type: integer
    type: object
  services.MonthlyTaxSummary:
    properties:
      month:
        type: integer
      month_name:
        type: string
      tax_amount:
        type: number
      year:
        type: integer
    type: object
...
  services.PurchaseItemDetail:
    properties:
      invoice_number:
        type: string
      product_code:
        type: string
      product_id:
        type: integer
      product_name:
        type: string
      purchase_date:
        type: string
      quantity:
        type: number
      total_price:
        type: number
      unit:
        type: string
      unit_price:
        type: number
    type: object
  services.PurchasePaymentAnalysis:
    properties:
      average_order_value:
        type: number
      cash_amount:
        type: number
      cash_percentage:
        type: number
      cash_purchases:
        type: integer
      credit_amount:
        type: number
      credit_percentage:
        type: number
      credit_purchases:
        type: integer
    type: object
  services.PurchaseReportData:
    properties:
      company:
        $ref: '#/definitions/services.CompanyInfo'
      completed_purchases:
        type: integer
      currency:
        type: string
      end_date:
        type: string
      generated_at:
        type: string
      outstanding_payables:
        type: number
      payment_analysis:
        $ref: '#/definitions/services.PurchasePaymentAnalysis'
      purchases_by_category:
        items:
          $ref: '#/definitions/services.CategoryPurchaseSummary'
        type: array
      purchases_by_month:
        items:
          $ref: '#/definitions/services.MonthlyPurchaseSummary'
        type: array
      purchases_by_vendor:
        items:
          $ref: '#/definitions/services.VendorPurchaseSummary'
        type: array
      start_date:
        type: string
      tax_analysis:
        $ref: '#/definitions/services.PurchaseTaxAnalysis'
      total_amount:
        type: number
      total_paid:
        type: number
      total_purchases:
        type: integer
    type: object
  services.PurchaseTaxAnalysis:
    properties:
      average_tax_rate:
        type: number
      tax_by_month:
        items:
          $ref: '#/definitions/services.MonthlyTaxSummary'
        type: array
      tax_reclaimable_amount:
        type: number
      total_tax_amount:
        type: number
      total_taxable_amount:
        type: number
    type: object
...
  services.VendorPurchaseSummary:
    properties:
      items:
        items:
          $ref: '#/definitions/services.PurchaseItemDetail'
        type: array
      last_purchase_date:
        type: string
      outstanding:
        type: number
      payment_method:
        type: string
      status:
        type: string
      total_amount:
        type: number
      total_paid:
        type: number
      total_purchases:
        type: integer
      vendor_id:
        type: integer
      vendor_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur
    lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sistema Akuntansi API
  version: "1.0"
paths:
...
  /api/monitoring/balance-sync:
    get:
      consumes:
      - application/json
      description: Check synchronization between cash/bank accounts and GL accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.BalanceMonitoringResult'
      security:
      - Bearer: []
      summary: Check balance synchronization
      tags:
      - Balance Monitoring
  /api/monitoring/discrepancies:
    get:
      consumes:
      - application/json
      description: Get list of current balance discrepancies between cash/bank and
        GL accounts
      parameters:
      - default: 50
        description: Limit number of results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.BalanceDiscrepancy'
            type: array
      security:
      - Bearer: []
      summary: Get current balance discrepancies
      tags:
      - Balance Monitoring
  /api/monitoring/fix-discrepancies:
    post:
      consumes:
      - application/json
      description: Automatically fix balance discrepancies by updating GL account
        balances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - Bearer: []
      summary: Fix balance discrepancies
      tags:
      - Balance Monitoring
  /api/monitoring/sync-status:
    get:
      consumes:
      - application/json
      description: Get a quick summary of balance synchronization status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - Bearer: []
      summary: Get synchronization status summary
      tags:
      - Balance Monitoring
...
  /api/v1/ssot-reports/purchase-report:
    get:
      description: Generate comprehensive purchase analysis using SSOT journal data
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - default: json
        description: Output format (json, pdf)
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.PurchaseReportData'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get purchase report integrated with SSOT journal
      tags:
      - SSOT Reports
...
  /api/v1/ssot-reports/purchase-summary:
    get:
      consumes:
      - application/json
      description: Get a quick summary of purchase report data
      parameters:
      - description: Start date (YYYY-MM-DD format)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD format)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Purchase summary data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/controllers.GetPurchaseReportResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.GetPurchaseReportResponse'
      summary: Get Purchase Report Summary
      tags:
      - Purchase Reports
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"